requirements.txt
bs4==0.0.1
python-dotenv==0.15.0
requests==2.23.0
schedule==0.6.0
twilio==6.51.0